Añadido Fade a algunos botones del Menu y Tutorial

+++ captain_brinetooth/captain_brinetooth/sdlutils/SDLUtils.cpp
@@ -31,7 +31,7 @@ SDLUtils::~SDLUtils() {
 }
 
 std::string SDLUtils::getNameFilePath(std::string& path){
-	int puntoCorte = path.find_last_of("\\");
+	size_t puntoCorte = path.find_last_of("\\");
 	int length = path.length(); //Final del string
 	std::string namePath = path.substr(puntoCorte + 1, length);
 	return namePath;
